SetupOsmMatcher failes currently, OsmMatcher.nodeCriteria need to be serialized correctly.

+++ src/main/java/org/projectzion/game/persitence/entities/Node.java
@@ -19,7 +19,7 @@ public class Node {
 
     private Long osmId;
 
-    private Long respawnTime;
+    private int respawnTime;
 
     @ManyToOne
     private NodeType nodeType;
